+++ src/components/Molecules/Membership/Membership.md
@@ -1,4 +1,4 @@
-Membership:
+# Membership
 
 ```js
 import data from './dev-data/data';
