+++ src/components/Molecules/Membership/MoneyBuy/MoneyBuy.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled, { css } from 'styled-components';
 
-import Input from '../../Atoms/Input/Input';
+import Input from '../../../Atoms/Input/Input';
 
 const MoneyBox = styled(Input)`
   display: block;
